ASoC: msm-cpe-lsm: Correct handling of stopping LAB thread
It is possible that the kthread used for LAB buffering could get stopped
even before it was scheduled. In such cases, the driver still waits for
the thread function to complete, which results into timeout and an error
sent to userspace. Fix handling of stopping LAB thread based on return
value of kthread_stop.

Bug: 24499363
Change-Id: I854d4407ced4dd3b50339614872ee12a9478280f
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: Banajit Goswami <bgoswami@codeaurora.org>
Signed-off-by: Mekala Natarajan <mekalan@codeaurora.org>